% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert}
\alias{assert}
\alias{assert_string}
\alias{assert_bool}
\alias{assert_scalar_int}
\title{Assertions}
\usage{
assert_string(x, allow_empty = TRUE)

assert_bool(x)

assert_scalar_int(x, integerish = FALSE)
}
\arguments{
\item{x}{The object to check}

\item{allow_empty}{Boolean. Should \code{assert_string()} allow an empty string?
The default is \code{TRUE}.}

\item{integerish}{Boolean. Should \code{assert_scalar_int()} allow \code{x} to be
integerish, rather than a strict \code{integer}?}
}
\value{
The input (or an error if assertion is failed)
}
\description{
\code{assert_string()} asserts that an object is a length 1 \code{character} vector

\code{assert_bool()} asserts that an object is a length 1 \code{logical} that is either
\code{TRUE} or \code{FALSE}

\code{assert_scalar_int()} asserts than an object is a scalar \code{integer}
(or integerish)
}
\keyword{internal}
