% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom.R
\name{booleans}
\alias{booleans}
\alias{is_phantom}
\alias{is_process}
\alias{is_ps_handle}
\title{Boolean Tests}
\usage{
is_phantom(x)

is_process(x)

is_ps_handle(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
\code{is_phantom()} tests whether an object is a PhantomJS list

\code{is_process()} tests whether an object is a \code{process}

\code{is_ps_handle()} tests whether an object is a \code{ps_handle}
}
\keyword{internal}
