% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs-loginout.R
\name{nbs-loginout}
\alias{nbs-loginout}
\alias{nbs_login}
\alias{nbs_logout}
\title{Log In and Out of NBS}
\usage{
nbs_login(session, creds = nbs_creds(), show_screen = rlang::is_interactive())

nbs_logout(session, show_screen = rlang::is_interactive())
}
\arguments{
\item{session}{A webdriver \code{\link[webdriver:Session]{Session}}}

\item{creds}{A \code{secret} containing NBS login credentials}

\item{show_screen}{\code{logical}. Should the browser screen be displayed on
return?}
}
\value{
A webdriver \code{\link[webdriver:Session]{Session}}
}
\description{
\code{nbs_login()} logs a user into NBS in a PhantomJS browsing session
}
\details{
\code{nbs_logout()} logs a user out of NBS in a PhantomJS browsing session
}
\keyword{internal}
