% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs-session.R
\name{nbs-session}
\alias{nbs-session}
\alias{nbs_start}
\alias{nbs_end}
\title{Start or End an NBS Browsing Session}
\usage{
nbs_start(phantom, creds = nbs_creds(), show_screen = rlang::is_interactive())

nbs_end(session, show_screen = rlang::is_interactive())
}
\arguments{
\item{phantom}{A list containing a PhantomJS \code{process} and \code{integer} port
number}

\item{creds}{A \code{secret} containing NBS login credentials}

\item{show_screen}{\code{logical}. Should the browser screen be displayed on
return?}

\item{session}{A webdriver \code{\link[webdriver:Session]{Session}}}
}
\value{
A webdriver \code{\link[webdriver:Session]{Session}}
}
\description{
\code{nbs_start()} starts a browsing session and logs into NBS
}
\details{
\code{nbs_end()} logs out of NBS and ends the browsing session
}
