% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_creds.R
\name{nbs_creds}
\alias{nbs_creds}
\alias{nbs_set_creds}
\title{NBS Credentials}
\usage{
nbs_creds(set_empty = TRUE)

nbs_set_creds(user = NULL)
}
\arguments{
\item{set_empty}{Boolean. Should \code{nbs_creds()} ask the user for their
credentials if none are found?}

\item{user}{String. NBS username; if empty, \code{set_nbs_creds()} will prompt the
user for their username.}
}
\value{
A \code{secret} containing the NBS username and password
}
\description{
\code{nbs_creds()} gets the username/password for NBS, if they exist. By default,
it asks the user to set the credentials if they do not exist.

\code{nbs_set_creds()} sets or resets the NBS username and password.
}
