% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs_search.R
\name{nbs_go_to_search}
\alias{nbs_go_to_search}
\title{Go to Advanced Search Tab}
\usage{
nbs_go_to_search(
  session,
  type = c("inv", "pat", "lab"),
  show_screen = rlang::is_interactive()
)
}
\arguments{
\item{session}{A \code{\link[webdriver:Session]{Session}} object}

\item{type}{The search entity (patient, investigation, or lab)}

\item{show_screen}{Should the browser screen be displayed on return?}
}
\value{
The \code{session} (invisibly)
}
\description{
Go to Advanced Search Tab
}
\keyword{internal}
