% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbscar-package.R
\docType{package}
\name{nbscar-package}
\alias{nbscar}
\alias{nbscar-package}
\title{nbscar: TN NBS Web Driver}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

nbscar allows interfacing with the TN NBS
    system from R.  It uses the webdriver package to programmatically
    operate the NBS UI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/nbscar}
  \item Report bugs at \url{https://github.com/jesse-smith/nbscar/issues}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{jesse.smith@shelbycountytn.gov}

}
\keyword{internal}
