% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{phtm-session}
\alias{phtm-session}
\alias{sesh_start}
\alias{sesh_end}
\title{Create and Delete a Browsing Session}
\usage{
sesh_start(phantom)

sesh_end(session)
}
\arguments{
\item{phantom}{A list containing the PhantomJS \code{process} and \code{integer} port
number}
}
\value{
A webdriver \code{\link[webdriver:Session]{Session}}
}
\description{
\code{sesh_start()} instantiates a browsing session using a headless PhantomJS
web driver

\code{sesh_end()} deletes a browsing session and finalizes the \code{Session} object
}
\keyword{internal}
