% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom.R
\name{phtm_find}
\alias{phtm_find}
\title{Find PhantomJS Processes Running in Background}
\usage{
phtm_find(parent = ps::ps_handle(), recursive = FALSE)
}
\arguments{
\item{parent}{The parent process to search}

\item{recursive}{Boolean. Should all descendants be returned, or only
children?}
}
\value{
A \code{list} of \code{ps_handle}s
}
\description{
Find PhantomJS Processes Running in Background
}
\keyword{internal}
