% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom.R
\name{phtm_kill}
\alias{phtm_kill}
\title{Kill PhantomJS Processes}
\usage{
phtm_kill(parent = ps::ps_handle(), excl = NULL, recursive = FALSE)
}
\arguments{
\item{parent}{The parent process of the PhantomJS processes to kill; the
default is the current R process}

\item{excl}{PhantomJS processes to keep alive. Either a list containing a
PhantomJS \code{process} object, a PhantomJS \code{process} object, a PhantomJS \code{ps}
object, or a list containing multiple of these objects, all of the same
type.}

\item{recursive}{Boolean. Should \code{phtm_kill()} kill all PhantomJS
descendants of \code{parent} or only children?}
}
\value{
An \code{integer} number of PhantomJS processes killed
}
\description{
\code{phtm_kill()} kills PhantomJS processes that are children (or descendants)
of a given parent process. This is useful for shutting down orphaned
PhantomJS processes.
}
