% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom.R
\name{phtm_ps}
\alias{phtm_ps}
\title{Extract the \code{ps_handle} of a PhantomJS list or \code{process}}
\usage{
phtm_ps(x)
}
\arguments{
\item{x}{A list with a PhantomJS \code{process} element named \code{process}, or a
\code{process} object}
}
\value{
A \code{ps_handle}
}
\description{
Extract the \code{ps_handle} of a PhantomJS list or \code{process}
}
\keyword{internal}
