% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom.R
\name{phtm_start}
\alias{phtm_start}
\title{Start a PhantomJS Process}
\usage{
phtm_start(timeout = 5, debug = c("info", "error", "warn", "debug"))
}
\arguments{
\item{timeout}{How long to wait (in seconds) for the connection to the
PhantomJS process to be established. Can be a decimal.}

\item{debug}{PhantomJS debug level}
}
\value{
A list containing the PhantomJS \code{process} and \code{integer} port number
}
\description{
\code{phtm_start()} starts a PhantomJS process that can be used to instantiate
headless web browsing sessions.
}
