% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-secret.R
\name{secret-helpers-internal}
\alias{secret-helpers-internal}
\alias{scrt_x}
\alias{scrt_secrets}
\alias{scrt_names}
\alias{scrt_set_names}
\alias{new_secret}
\alias{expand_secrets}
\title{Internal \code{secret} Helpers}
\usage{
scrt_x(x)

scrt_secrets(x)

scrt_names(x)

scrt_set_names(x, names)

new_secret(x, secrets = NULL, locked = TRUE, class = NULL, ...)

expand_secrets(secrets, x)
}
\arguments{
\item{x}{A \code{character} vector}

\item{secrets}{\code{logical} of the same size as \code{x}. Elements to hide when
printing (defaults to all)}

\item{class}{string. An optional subclass}

\item{...}{Additional attributes to give to the subclass}
}
\value{
A \code{secret}
}
\description{
Internal \code{secret} Helpers
}
\keyword{internal}
