% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-secret.R
\name{secret-methods}
\alias{secret-methods}
\alias{format.nbscar_secret}
\alias{obj_str_footer.nbscar_secret}
\alias{names.nbscar_secret}
\alias{names<-.nbscar_secret}
\alias{vec_ptype_abbr.nbscar_secret}
\alias{vec_ptype_full.nbscar_secret}
\alias{vec_ptype2.nbscar_secret.nbscar_secret}
\alias{vec_ptype2.nbscar_secret.character}
\alias{vec_ptype2.character.nbscar_secret}
\alias{vec_cast.nbscar_secret.nbscar_secret}
\alias{vec_cast.nbscar_secret.character}
\alias{vec_cast.character.nbscar_secret}
\alias{$.nbscar_secret}
\alias{[[.nbscar_secret}
\title{S3 Methods for \code{secret} Objects}
\usage{
\method{format}{nbscar_secret}(x, ...)

\method{obj_str_footer}{nbscar_secret}(x, ...)

\method{names}{nbscar_secret}(x)

\method{names}{nbscar_secret}(x) <- value

\method{vec_ptype_abbr}{nbscar_secret}(x, ...)

\method{vec_ptype_full}{nbscar_secret}(x, ...)

\method{vec_ptype2}{nbscar_secret.nbscar_secret}(x, y, ...)

\method{vec_ptype2}{nbscar_secret.character}(x, y, ...)

\method{vec_ptype2}{character.nbscar_secret}(x, y, ...)

\method{vec_cast}{nbscar_secret.nbscar_secret}(x, to, ...)

\method{vec_cast}{nbscar_secret.character}(x, to, ...)

\method{vec_cast}{character.nbscar_secret}(x, to, ...)

\method{$}{nbscar_secret}(x, i)

\method{[[}{nbscar_secret}(x, i)
}
\description{
S3 Methods for \code{secret} Objects
}
\keyword{internal}
