% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-secret.R
\name{secret}
\alias{secret}
\alias{as_secret}
\alias{is_secret}
\alias{scrt_lock}
\alias{scrt_unlock}
\alias{scrt_set_secrets}
\alias{scrt_locked}
\title{Create a \code{secret}}
\usage{
secret(x = character(), secrets = NULL, locked = TRUE)

as_secret(x, secrets = NULL)

is_secret(x)

scrt_lock(x)

scrt_unlock(x)

scrt_set_secrets(x, secrets)

scrt_locked(x)
}
\arguments{
\item{x}{A \code{character} vector}

\item{secrets}{\code{logical} of the same size as \code{x}. Elements to hide when
printing (defaults to all)}
}
\value{
A \code{secret}
}
\description{
\code{secret}s are character vectors that print with the specified \code{secrets}
hidden by asterisks (\verb{***}).
}
\details{
\code{secret()} and \code{as_secret()} creates a \code{secret}, casting to \code{character} first
if needed and possible

\code{is_secret()} tests whether an object is a well-formed \code{secret}

\code{scrt_lock()} and \code{scrt_unlock()} lock and unlock the \code{secrets}

\code{scrt_set_secrets()} sets secret values in a \code{secret}

\code{scrt_locked()}  tests whether a \code{secret} is locked or unlocked
}
\keyword{internal}
