% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std_class}
\alias{std_class}
\title{Standardize a Class}
\usage{
std_class(...)
}
\arguments{
\item{...}{\code{character} vectors of class names}
}
\description{
Combine class names and remove missing values
}
