% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-secret.R
\name{validate_secret}
\alias{validate_secret}
\title{Validate a Well-Formed \code{secret} Object}
\usage{
validate_secret(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
\code{x} (or throw an error if invalid)
}
\description{
Validate a Well-Formed \code{secret} Object
}
\keyword{internal}
