% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_rstudio}
\alias{verify_rstudio}
\title{Verify Whether RStudio is Available Via API}
\usage{
verify_rstudio(version_needed = "0.99.853")
}
\arguments{
\item{version_needed}{Optional minimal RStudio version}
}
\value{
\code{TRUE} if the rstudioapi package is installed and RStudio is running;
otherwise, \code{FALSE}
}
\description{
Verify Whether RStudio is Available Via API
}
\keyword{internal}
